refactor(errors): split 404 and 500 handlers and add registerErrorHandlers
- Extracted global error handler (500) into a dedicated function
- Extracted not found handler (404) into a dedicated function
- Added registerErrorHandlers utility to apply both middlewares in a clean way
- Simplified interfaces (removed logger dependency)
feat(server): move server initialization to dedicated package

- Introduced reusable server bootstrap function
- Encapsulated express app creation and middleware registration
- Prepared foundation for reusing the same init logic across microservices
