+++ packages/common/src/index.ts
@@ -1,9 +1,8 @@
 export { requestLoggerDev } from './requestLoggerDev.ts';
 export { setupCommonMiddleware } from './setupCommonMiddleware.ts'; 
-export { setupErrorHandling } from './setupErrorHandling.ts';
 export { globalRateLimiter } from './rate-limiters.ts';
-export { globalErrorHandler } from './middleware/globalErrorHandler.ts';
-export { notFoundHandler } from './middleware/notFoundHandler.ts';
+export { globalErrorHandler } from './server-utils/middleware/globalErrorHandler.ts';
+export { notFoundHandler } from './server-utils/middleware/notFoundHandler.ts';
 export { default as logger} from './utils/logger.ts';
 export * from './utils/server/shutdown.ts';
 export { healthRouter } from  './router/health/health.route.ts';
@@ -12,3 +11,4 @@ export { COMMON_CONFIG, CommonEnv } from './config/env.ts';
 // re-export UUID
 export { v4 as uuidv4, v5 as uuidv5 } from 'uuid';
 export { setupServer } from './server-utils/index.ts';
+export { registerErrorHandlers } from './server-utils/middleware/registerErrorHandlers.ts';
