+++ services/auth-service/src/index.ts
@@ -1,9 +1,10 @@
 import 'dotenv/config';
-import { setupServer } from '@zerodayz7/common';
+import { registerErrorHandlers, setupServer } from '@zerodayz7/common';
 import { logger } from '@zerodayz7/common';
-import app from './app.js';
 import env from './config/env.js';
-
+import { createApp } from './setup/setupInit.ts';
+import { setupRoutes } from './setup/setupRoutes.ts';
+ 
 const serverConfig = {
   port: env.PORT,
   name: env.NAME,
@@ -13,9 +14,16 @@ const serverConfig = {
   requestTimeout: env.REQUEST_TIMEOUT,
 };
 
+const app = createApp();
+
 async function initializeServices(): Promise<void> {
   // Tutaj inicjalizacja specyficzna dla serwisu, np. Redis, session manager
-  logger.info('Inicjalizacja specyficznych serwisów dla mikeoservisu');
+  logger.info('🔧 Inicjalizacja specyficznych serwisów dla mikeoservisu');
+  setupRoutes(app);
+  registerErrorHandlers(app, {
+    serviceName: serverConfig.name,
+    isDev: serverConfig.nodeEnv === 'development',
+  });
   // np. await redis.connect();
 }
