+++ services/auth-service/src/setup/setupInit.ts
@@ -1,0 +1,7 @@
+import { logger, setupCommonMiddleware } from '@zerodayz7/common';
+import { corsOptions } from '../config/cors.config.ts';
+
+export function createApp() {
+  logger.info('🛠️  Creating Express app...');
+  return setupCommonMiddleware({ cors: corsOptions });
+}
